requirements.txt
httpx>=0.24.0
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.0.0
